[loginmanager] Initialize nsLoginManager only after profile is accessible
At least since Gecko38 nsLoginManager tries to initialize storage is a
part of its initialization procedure. But the storage can't be
initialized until the user profile is available for a component.
It's a wrong assumption that the profile is available upon the
"app-startup" notification. In fact the availability of the profile
is announced with the "profile-after-change" notification.