[embedlite-components] Update input zooming to work with Gecko 38. Contributes to JB#32818
Calculate virtual keyboard open state with a helper function
that takes into account the space that is reserved by the vkb.

ViewportHandler is copied from Android FF browser.js. With help
of that we can check if zooming is really allowed. In addition, we
don't do zooming if the calculated viewport is the same as the
current composited css rect.